Remove unused CtrlDispatchApiEvent() dependency
This commit removes the CtrlDispatchApiEvent() dependency from the
binding. The feature was unused as no events are processed this way in
the code.

The API is also no more available on the controller side, which was
making it impossible to start the binding with the latest controller
sources (because of undefined symbols errors).