fixed the deflatten logics
There were caveats around objects in arrays and and array in objects
The code has been significantly simplifyed by the usage of jsoncpp
in that function.
Notice that the deflatten mechanism is only used by ts_mget,
because, for performances reasons, the flatten strings
are kept as such with ts_mrange.

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>