Revert "[cmake] Change afb-helpers to afb-libhelpers"
This reverts commit 154f796f77bcf1ac8a3f0c32f24a846491ac3177.

The library dependency was changed from afb-helpers to afb-libhelpers,
at the time seemingly for packaging reasons.

The problem is that the latter is version 4 of the library, whose behavior
differs from v3, for instance in the GetAFBRootDirPath() routine. This
was causing crashes when the binding would be loaded: as it uses the
controller library APIs, it would end up calling LuaConfigLoad() which
would fault because GetAFBRootDirPath() returns NULL in v4.

A potential fix in LuaConfigLoad() will be the subject of a separate
commit.